// src/app/pages/06-produccion/interfaces/proceso_fabricacion.interface.ts
export interface ProduccionProcesoFabricacionInterface {
    IdProcesoFabricacion: number;
    IdProductoTerminado: number;
    CodigoMaterial: string;
    NombreMaterial: string;
    AnchoMaterial: number;
    CodigoMaquina1: number;
    NombreMaquina1: string;
    CodigoMaquina2: number;
    NombreMaquina2: string;
    CodigoMaquina3: number;
    NombreMaquina3: string;
    CodigoMaquina4: number;
    NombreMaquina4: string;
    CodigoMaquina5: number;
    NombreMaquina5: string;
    CodigoMaquina6: number;
    NombreMaquina6: string;
}